--- plugins/dev/1761.go
@@ -1,39 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("1761", "Digital Journal", "https://www.digitaljournal.com/")
-
-	engine.SetStartingURLs([]string{"https://www.digitaljournal.com/pr/news/index.html"})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        true,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-
-	engine.OnHTML(".h5", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.News)
-	})
-
-	engine.OnHTML(".col-lg-8", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-
-	engine.OnHTML(".d-flex > div:nth-last-child(1) > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.Index)
-	})
-}
