--- plugins/dev/1762.go
@@ -1,42 +1,0 @@
-package dev
-
-import (
-	"strings"
-
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("1762", "镜子晚报", "https://weeklysamirror.news/")
-
-	engine.SetStartingURLs([]string{"https://weeklysamirror.news/wp-sitemap.xml"})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        true,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-
-	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
-		switch {
-		case strings.Contains(element.Text, "wp-sitemap-posts-post"):
-			engine.Visit(element.Text, crawlers.Index)
-		default:
-			engine.Visit(element.Text, crawlers.News)
-		}
-	})
-
-	engine.OnHTML(".entry-content", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-}
