--- plugins/dev/1767.go
@@ -1,43 +1,0 @@
-package dev
-
-import (
-	"megaCrawler/crawlers"
-	"megaCrawler/extractors"
-
-	"github.com/gocolly/colly/v2"
-)
-
-func init() {
-	engine := crawlers.Register("1767", "周日快报", "https://www.scottishdailyexpress.co.uk/")
-
-	engine.SetStartingURLs([]string{
-		"https://www.scottishdailyexpress.co.uk/news/scottish-news/",
-		"https://www.scottishdailyexpress.co.uk/news/uk-news/",
-		"https://www.scottishdailyexpress.co.uk/news/world-news/",
-	})
-
-	extractorConfig := extractors.Config{
-		Author:       true,
-		Image:        true,
-		Language:     true,
-		PublishDate:  true,
-		Tags:         true,
-		Text:         false,
-		Title:        true,
-		TextLanguage: "",
-	}
-
-	extractorConfig.Apply(engine)
-
-	engine.OnHTML(".headline", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.News)
-	})
-
-	engine.OnHTML(".content-column", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
-	})
-
-	engine.OnHTML(".pagi-next > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Visit(element.Attr("href"), crawlers.Index)
-	})
-}
