+++ plugins/storage/1762.go
@@ -1,4 +1,4 @@
-package dev
+package storage
 
 import (
 	"strings"
@@ -10,9 +10,9 @@ import (
 )
 
 func init() {
-	engine := crawlers.Register("1773", "巴基斯坦官方新闻", "https://en.dailypakistan.com.pk/")
+	engine := crawlers.Register("1762", "镜子晚报", "https://weeklysamirror.news/")
 
-	engine.SetStartingURLs([]string{"https://en.dailypakistan.com.pk/sitemap_index.xml"})
+	engine.SetStartingURLs([]string{"https://weeklysamirror.news/wp-sitemap.xml"})
 
 	extractorConfig := extractors.Config{
 		Author:       true,
@@ -20,7 +20,7 @@ func init() {
 		Language:     true,
 		PublishDate:  true,
 		Tags:         true,
-		Text:         true,
+		Text:         false,
 		Title:        true,
 		TextLanguage: "",
 	}
@@ -29,10 +29,14 @@ func init() {
 
 	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
 		switch {
-		case strings.Contains(element.Text, "post-sitemap"):
+		case strings.Contains(element.Text, "wp-sitemap-posts-post"):
 			engine.Visit(element.Text, crawlers.Index)
-		case !strings.Contains(element.Request.URL.String(), "sitemap_index.xml"):
+		default:
 			engine.Visit(element.Text, crawlers.News)
 		}
+	})
+
+	engine.OnHTML(".entry-content", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
 	})
 }
