+++ plugins/storage/1763.go
@@ -1,0 +1,45 @@
+package storage
+
+import (
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("1763", "每日金融时报", "https://www.ft.lk/")
+
+	engine.SetStartingURLs([]string{
+		"https://www.ft.lk/top-story/26",
+		"https://www.ft.lk/business-news/34",
+		"https://www.ft.lk/news/56",
+		"https://www.ft.lk/news/44",
+		"https://www.ft.lk/financial-services/42",
+	})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        true,
+		Language:     true,
+		PublishDate:  true,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+
+	engine.OnHTML(".date", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Visit(element.Attr("href"), crawlers.News)
+	})
+
+	engine.OnHTML(".inner-content", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+
+	engine.OnHTML(".active + li > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Visit(element.Attr("href"), crawlers.Index)
+	})
+}
