+++ plugins/storage/1769.go
@@ -1,0 +1,39 @@
+package storage
+
+import (
+	"megaCrawler/crawlers"
+	"megaCrawler/extractors"
+
+	"github.com/gocolly/colly/v2"
+)
+
+func init() {
+	engine := crawlers.Register("1769", "爱尔兰观察报", "https://www.irishexaminer.com/")
+
+	engine.SetStartingURLs([]string{"https://www.irishexaminer.com/search/?cosearch=News"})
+
+	extractorConfig := extractors.Config{
+		Author:       true,
+		Image:        true,
+		Language:     true,
+		PublishDate:  true,
+		Tags:         true,
+		Text:         false,
+		Title:        true,
+		TextLanguage: "",
+	}
+
+	extractorConfig.Apply(engine)
+
+	engine.OnHTML(".card > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Visit(element.Attr("href"), crawlers.News)
+	})
+
+	engine.OnHTML(".nfyc-detail-view", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Content += element.Text
+	})
+
+	engine.OnHTML(".pagination > li:nth-last-child(2) > a", func(element *colly.HTMLElement, ctx *crawlers.Context) {
+		ctx.Visit(element.Attr("href"), crawlers.Index)
+	})
+}
