+++ plugins/storage/1772.go
@@ -1,18 +1,17 @@
-package dev
+package storage
 
 import (
-	"strings"
-
 	"megaCrawler/crawlers"
 	"megaCrawler/extractors"
+	"strings"
 
 	"github.com/gocolly/colly/v2"
 )
 
 func init() {
-	engine := crawlers.Register("1765", "股票代码报告", "https://www.tickerreport.com/")
+	engine := crawlers.Register("1772", "观察家报", "https://www.spectator.co.uk/")
 
-	engine.SetStartingURLs([]string{"https://www.tickerreport.com/sitemap.xml"})
+	engine.SetStartingURLs([]string{"https://www.spectator.co.uk/sitemap_index.xml"})
 
 	extractorConfig := extractors.Config{
 		Author:       true,
@@ -20,7 +19,7 @@ func init() {
 		Language:     true,
 		PublishDate:  true,
 		Tags:         true,
-		Text:         false,
+		Text:         true,
 		Title:        true,
 		TextLanguage: "",
 	}
@@ -28,7 +27,7 @@ func init() {
 	extractorConfig.Apply(engine)
 
 	engine.OnXML("//loc", func(element *colly.XMLElement, ctx *crawlers.Context) {
-		if strings.Contains(element.Request.URL.String(), "sitemap.xml") {
+		if strings.Contains(element.Request.URL.String(), "sitemap_index") {
 			if strings.Contains(element.Text, "post") {
 				ctx.Visit(element.Text, crawlers.Index)
 				return
@@ -36,9 +35,5 @@ func init() {
 			return
 		}
 		ctx.Visit(element.Text, crawlers.News)
-	})
-
-	engine.OnHTML(".entry", func(element *colly.HTMLElement, ctx *crawlers.Context) {
-		ctx.Content += element.Text
 	})
 }
